split websockets module
